Sensor Circuit Integrated
Sensor circuit and ADC code has been added to the slave function. Slaves
can now be initialized with their address, and participate in a practice
run. Rasp pi tells the first panel it is active, and once its sensor is
triggered, second panel becomes active and waits for sensor trigger.